feat: add sweep on and off function

+++ src/main.py
@@ -19,8 +19,10 @@ def set_seed(seed):
     torch.cuda.manual_seed(seed)  # GPU 시드 고정
     torch.cuda.manual_seed_all(seed)  # 모든 GPU에 대한 시드 고정 (멀티 GPU 환경)
 
+
 def load_and_merge_config(base_config, wandb_config):
     return {**base_config, **wandb_config}
+
 
 def get_trainer(config, log_name):
     return pl.Trainer(
@@ -32,7 +34,8 @@ def get_trainer(config, log_name):
                  pl.loggers.WandbLogger()]
                     )
  
-def train(base_config):
+ 
+def train(base_config, model_path):
     
     # wandb 설정
     with wandb.init(config = base_config) as run:
@@ -40,17 +43,15 @@ def train(base_config):
         # dataloader와 model을 생성
         dataloader = Dataloader(config)
         
-        if model_path and os.path.exists(model_path):
-            print("저장된 모델을 확인하였으며 해당 모델을 load하여 학습합니다")
+        if os.path.exists(model_path):
             print(f"Loading model from {model_path}")
+            model = Model(config)
             model = torch.load(model_path)
+            
         else:
-            print("model_path를 확인하지 못하였으며 config.yaml에 있는 모델명으로 새로 학습합니다")
             print("Creating new model")
             model = Model(config)
-        
-        model = Model(config)
-        
+    
         # logging name 설정
         log_name = f"{config['model']['name']}_bs{config['training']['batch_size']}_lr{config['training']['learning_rate']}"
     
@@ -61,8 +62,12 @@ def train(base_config):
         trainer.fit(model=model, datamodule=dataloader)
         trainer.test(model=model, datamodule=dataloader)
         
-        # 학습이 완료된 모델을 저장함.
-        torch.save(model, f"model_{wandb.run.id}.pt")
+        
+        
+        if not os.path.exists(model_path):
+            torch.save(model, model_path)
+        else:
+            torch.save(model, f"model_{wandb.run.id}.pt")
         
 
 def main():
@@ -70,37 +75,47 @@ def main():
     parser = argparse.ArgumentParser()
     parser.add_argument('--mode', type=str, default='train', help='train or inference mode')
     parser.add_argument('--model_path', type=str, default ='model.pt', help='path to the model weights file')
+    parser.add_argument('--sweep', type=bool, default=False, help='whether to use wandb sweep')
+    
     args = parser.parse_args()
         
     # 재현을 위한 seed 고정
     set_seed(base_config["seed"])
     
+    wandb.login(key="aadeea1600199a35fa358306a6e7c09a4240f709")
+
     if args.mode == 'train':
         print("Running on train mode")
         
-        sweep_config = {
-            'method': 'random',
-            'metric': {'name': 'val_loss', 'goal': 'minimize'},
-            'parameters': {
-                'training.batch_size': {'values': [4, 8, 16]},
-                'training.learning_rate': {'min': 1e-5, 'max': 1e-3},
-                'training.epochs': {'values': [3, 5, 10, 20]},
-                'training.weight_decay': {'min': 0.001, 'max': 0.1},
-                'training.optimizer': {'values': ['AdamW', 'Adam', 'SGD']},
-                'training.scheduler': {'values': ['CosineAnnealingLR', 'StepLR', 'ReduceLROnPlateau']},
-                'training.loss': {'values': ['L1Loss', 'L2Loss']},
-                'data.dropout_rate': {'min': 0.1, 'max': 0.5},
-                'data.augmentation.0.params.probability': {'min': 0.5, 'max': 1.0}
+        if args.sweep:
+            print("Running Sweep")
+            sweep_config = {
+                'method': 'random',
+                'metric': {'name': 'val_loss', 'goal': 'minimize'},
+                'parameters': {
+                    'training.batch_size': {'values': [4, 8, 16]},
+                    'training.learning_rate': {'min': 1e-5, 'max': 1e-3},
+                    'training.epochs': {'values': [3, 5, 10, 20]},
+                    'training.weight_decay': {'min': 0.001, 'max': 0.1},
+                    'training.optimizer': {'values': ['AdamW', 'Adam', 'SGD']},
+                    'training.scheduler': {'values': ['CosineAnnealingLR', 'StepLR', 'ReduceLROnPlateau']},
+                    'training.loss': {'values': ['L1Loss', 'L2Loss']},
+                    'data.dropout_rate': {'min': 0.1, 'max': 0.5},
+                    'data.augmentation.0.params.probability': {'min': 0.5, 'max': 1.0}
+                }
             }
-        }
         
-        wandb.login(key="aadeea1600199a35fa358306a6e7c09a4240f709")
-        sweep_id = wandb.sweep(sweep_config, project="U-4-do", entity='nlp-01')
-        wandb.agent(sweep_id, lambda: train(base_config), count = 5) # 5번 실험 실행
+            sweep_id = wandb.sweep(sweep_config, project="U-4-do", entity='nlp-01')
+            wandb.agent(sweep_id, lambda: train(base_config, args.model_path), count = 5) # 5번 실험 실행
+        
+        else:
+            print("Running without wandb sweep")
+            train(base_config, args.model_path)
+            
        
     elif args.mode == 'inference':
         
-        # 디버깅 코드 추가 agr에 inference 입력시 이하 코드 출력
+        # 디버깅 코드 추가 mode에 inference 입력시 이하 코드 출력
         print('Running on Inference Mode')
         config = load_config()
         dataloader = Dataloader(config)
@@ -112,10 +127,6 @@ def main():
 
         # 예측된 결과를 형식에 맞게 반올림하여 준비합니다.
         predictions = list(round(float(i), 1) for i in torch.cat(predictions))
-
-        # output 형식을 불러와서 예측된 결과로 바꿔주고, output.csv로 출력합니다.
-        # output = pd.read_csv('./data/sample_submission.csv')
-        # output['target'] = predictions
         
         output_format = pd.read_csv('./data/sample_submission.csv')
         output = pd.DataFrame(columns=output_format.columns) # output 형식의 columns만 참고
